+++ src/componets/boards/todo/Todo.scss
@@ -1,0 +1,26 @@
+@import "src/style/Global";
+
+.todo {
+  background-color: $Contrast_Theme;
+  color: $Dark_Theme;
+  margin: 0.5rem 0.5rem;
+  border-top: 0.05rem solid $Medium_Theme;
+  border-bottom: 0.05rem solid $Medium_Theme;
+  border-radius: $Border_Application;
+  width: 90%;
+
+  .navTodo {
+    display: flex;
+    justify-content: center;
+
+    svg {
+      color: $Dark_Theme;
+      border: 0.05rem solid $Dark_Theme;
+      border-radius: $Border_Application;
+      width: 2.5rem;
+      height: 1.5rem;
+    }
+  }
+
+}
+
